@@ -31,6 +31,7 @@ environment:
 dependencies:
   animate_do: ^3.0.2
   animated_bottom_navigation_bar: ^1.2.0
+  animated_rail: ^2.0.2
   animations: ^2.0.7
   auto_size_text: ^3.0.0
   avatars: ^3.0.2
@@ -47,7 +48,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   datetime_picker_formfield_new: ^2.1.0
   draggable_fab: ^0.1.4
-  dropdown_button2: ^1.9.2
+  dropdown_button2: ^2.1.3
   equatable: ^2.0.5
   expansion_tile_card: ^3.0.0
   fancy_password_field: ^2.0.0
@@ -64,7 +65,7 @@ dependencies:
   flutter_svg: ^2.0.3
   flutter_zoom_drawer: ^3.0.4+1
   fluttertoast: ^8.2.1
-  font_awesome_flutter: ^10.1.0
+  font_awesome_flutter: ^10.4.0
   foundation_flutter: ^0.2.0
   geolocator: ^9.0.2
   getwidget: ^3.0.1
@@ -170,6 +171,8 @@ flutter:
     - assets/Image/sunny.png
     - assets/Image/wind.png
     - assets/Image/logo.png
+    - assets/new Updated images/megaphone.png
+    - assets/new Updated images/Notifications.png
     - null
   # To add assets to your application, add an assets section, like this:
   # assets:
