@@ -6,12 +6,10 @@ import 'package:intellensense/Pages/Notificationpages/Twitter.dart';
 import 'package:intellensense/Pages/Notificationpages/Youtube.dart';
 
 import '../components/sidebarPages/NewsTrsScreen.dart';
-import 'DrawerScreens/CandidatureAnalysis/NewsChannelSentiment.dart';
 import 'Notificationpages/NewsChannel.dart';
 import 'Notificationpages/Newspaper.dart';
 
 class Notifications extends StatefulWidget {
-
   @override
   State<Notifications> createState() => _NotificationsState();
 }
@@ -21,9 +19,7 @@ class _NotificationsState extends State<Notifications> {
   var _selectedIndex = 0;
   @override
   Widget build(BuildContext context) {
-    return
-    
-     Scaffold(
+    return Scaffold(
       backgroundColor: Color(0xffd2dfff),
       appBar: AppBar(
         elevation: 0,
@@ -44,7 +40,16 @@ class _NotificationsState extends State<Notifications> {
           height: 50,
         ),
       ),
-      bottomNavigationBar: BottomNavigationBar(selectedItemColor: Colors.black,
+      bottomNavigationBar: BottomNavyBar(
+        backgroundColor: Colors.blue.shade100,
+        selectedIndex: _selectedIndex,
+        showElevation: true, // use this to remove appBar's elevation
+        onItemSelected: (index) => setState(() {
+          _selectedIndex = index;
+          _pageController.animateToPage(index,
+              duration: Duration(milliseconds: 300), curve: Curves.ease);
+        }),
+        /*selectedItemColor: Colors.black,
        // backgroundColor: Color(0xffd2dfff),
         currentIndex: _selectedIndex,
         //showElevation: true, // use this to remove appBar's elevation
@@ -52,26 +57,24 @@ class _NotificationsState extends State<Notifications> {
           _selectedIndex = index;
           _pageController.animateToPage(index,
               duration: Duration(milliseconds: 300), curve: Curves.ease);
-        }),
+        }),*/
         items: [
-          BottomNavigationBarItem(
+          BottomNavyBarItem(
               icon: Image.asset(
                 'assets/icons/Social-Media-Icons-IS-10.png',
                 height: 25,
               ),
-              label: 'YouTube',
-              backgroundColor: Color(0xffd2dfff)
-          ),
-          BottomNavigationBarItem(
+              title: Text('YouTube'),
+        activeColor: Colors.blue),
+          BottomNavyBarItem(
               icon: Image.asset(
                 'assets/icons/fb.png',
                 height: 25,
               ),
-              label: 'FaceBook',
-              backgroundColor: Color(0xffd2dfff)
+              title: Text('FaceBook'),
+    activeColor: Colors.blue),
               //backgroundColor: Colors.blue
-          ),
-          BottomNavigationBarItem(
+          /*BottomNavigationBarItem(
               icon: Image.asset(
                 'assets/icons/Social-Media-Icons-IS-07.png',
                 height: 25,
@@ -79,34 +82,32 @@ class _NotificationsState extends State<Notifications> {
               label: 'Instagram',
               backgroundColor: Color(0xffd2dfff)
               //backgroundColor: Colors.blue
-          ),
-          BottomNavigationBarItem(
+          )*/
+          BottomNavyBarItem(
               icon: Image.asset(
                 'assets/new Updated images/intellisensesolutions-Icons-83.png',
                 height: 25,
               ),
-              label: 'Newspaper',
-              backgroundColor: Color(0xffd2dfff)
-              //backgroundColor: Colors.blue
-          ),
-          BottomNavigationBarItem(
-            icon: Image.asset(
-              'assets/icons/Social-Media-Icons-IS-08.png',
-              height: 25,
-            ),
-            label: 'Twitter',
-              backgroundColor: Color(0xffd2dfff)
-            //backgroundColor: Colors.blue,
-          ),
-          BottomNavigationBarItem(
-            icon: Image.asset(
-              'assets/new Updated images/news-71.png',
-              height: 25,
-            ),
-            label: 'NewsChannel',
-              backgroundColor: Color(0xffd2dfff)
-            //backgroundColor: Colors.blue,
-          ),
+              title: Text('Newspaper'),
+              activeColor: Colors.blue),
+          BottomNavyBarItem(
+              icon: Image.asset(
+                'assets/icons/Social-Media-Icons-IS-08.png',
+                height: 25,
+              ),
+            title: Text('Twitter'),
+            activeColor: Colors.blue,
+              //backgroundColor: Colors.blue,
+              ),
+          BottomNavyBarItem(
+              icon: Image.asset(
+                'assets/new Updated images/news-71.png',
+                height: 25,
+              ),
+              title: Text('NewsChannel'),
+            activeColor: Colors.blue,
+              //backgroundColor: Colors.blue,
+              ),
         ],
       ),
       body: Column(children: [
@@ -117,7 +118,7 @@ class _NotificationsState extends State<Notifications> {
               children: [
                 Youtube(),
                 Facebook(),
-                Instagram(),
+                //Instagram(),
                 Newspaper(),
                 Twitter(),
                 NewsChannelPage(),
